+++ PJ3T3_Postie/Core/Root/ContentView.swift
@@ -7,6 +7,12 @@
 
 import SwiftUI
 
+struct TestView: View {
+    var body: some View {
+        NaverMap()
+    }
+}
+
 struct ContentView: View {
     //ViewModels
     @ObservedObject var authViewModel = AuthViewModel.shared
