+++ Podfile
@@ -1,5 +1,5 @@
 # Uncomment the next line to define a global platform for your project
-# platform :ios, '9.0'
+platform :ios, '13.0'
 
 target 'PJ3T3_Postie' do
   # Comment the next line if you don't want to use dynamic frameworks
